handle old routes + edit fabfile to avoid always upgrading modules

+++ fabfile.py
@@ -12,7 +12,8 @@ def deploy(c):
         c.run("git stash")
         c.run("git checkout main")
         c.run("git pull")
-        c.run("pip install -r requirements.txt --upgrade")
+        if click.confirm("Upgrade modules ?", default=False):
+            c.run("pip install -r requirements.txt --upgrade")
         c.run("systemctl restart www.tahm-ken.ch_gunicorn.service")
